Make illegal items unusable
